api_yamdb/requirements.txt
asgiref==3.5.2
requests==2.26.0
django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0